backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LangGraph + LangChain
langgraph==0.2.60
langchain==0.3.17
langchain-openai==0.3.4
langchain-core==0.3.40

# Database
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Pydantic + settings
pydantic==2.10.5
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1
tenacity==9.0.0

# Stripe
stripe==11.4.1

# Observability
structlog==25.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
prometheus-client>=0.20.0
sentry-sdk[fastapi]==2.20.0

# Security
# python-jose removed: CVE-2024-33664 (algorithm confusion attack)
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4

# Rate limiting (H-06)
slowapi>=0.1.9
limits>=3.6.0

# Utilities
python-dotenv==1.0.1
orjson==3.10.14
ulid-py==1.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
factory-boy==3.3.1
